requirements.txt
streamlit==1.25.0
unstructured==0.10.19
PyPDF2==3.0.1
python-docx==1.1.2
pymongo==4.6.1
presidio-analyzer==2.2.33
spacy==3.7.2
crewai==0.28.8
langchain-google-genai==1.0.7
python-dotenv==1.0.0
setuptools<81